The outputs of an IR node can now be *.
Previously, the outputs of the last IR node of a block were simply ignored.
In cases where the number of expected values was not known, we set the
outputs to an empty list.  However, doing so has the disadvantage that we
cannot distinguish between a node that truly has zero outputs, and one that
may have outputs but happens to appear as the last node of a block whose
number of expected values is not known.